Adding transforms to documentations. Moving augmentations in transforms.
